requirements.txt
# Web framework
streamlit>=1.28.0

# Database connectors
redshift-connector>=2.0.918
psycopg2-binary>=2.9.9  # Alternative connector

# LLM integration
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.6

# Data handling
pandas>=1.5.0
numpy>=1.24.0

# Environment and utilities
python-dotenv>=1.0.0
sqlparse>=0.4.4  # For SQL validation
plotly>=5.17.0  # For data visualization

# Optional: for enhanced UI
streamlit-option-menu>=0.3.6 